requirements.txt
streamlit
crewai
crewai-tools
python-dotenv
langchain-openai
pyyaml
requests
